Rename files #JC19-329
