+++ apps/vanilla/src/constants/event.ts
